Create GEEcodeS2
This GEE script uses S2 images to create a nice mosaic for illustration. The resulting image is in 8bit and clouds are removed (if a suitable time range was used). Furthermore, different land classes are radiometrically corrected to achieve a better looking result (e.g. the brightness of snow areas is reduced)